requirements.txt
kafka-python
avro
elasticsearch
